Hackity hacky to skip running the fixers if a conflict marker shows up.
When debugging, I like writing lines of code like this:

    import pdb; pdb.set_trace() #<<<

It's annoying when `black` goes in and formats this, because it's more
work to remove later.

Not sure if I'll be able to upstream this to the ale folks, they may
feel it's a problem that needs to be fixed in the relevant fixer. For
now, I'm just planning to play around with this locally and see how it
goes.